Secret label selector configuration
Add a label selector configuration point to restrict the secrets picked up by this plugin. The syntax is based on the Kubernetes label selector api, https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#label-selectors

The use case for this is when there are multiple jenkins instance in a single namespace and you want to limit which secrets are available to
each instance.  For example, you might have some common secrets that are
applicable to both instances and others that should only be visibile to subset of
instances. With the label selector you can accomplish this by adding a selector
like so:

```yaml
# devops-jenkins-deployment.yaml
env:
  - name: JAVA_OPTS
     value: -Dcom.cloudbees.jenkins.plugins.kubernetes_credentials_provider.KubernetesCredentialProvider.labelSelector="mycompany.com/jenkins in (common devops)"
```

The secrets would then be configured with the selector labels
```yaml
# deveops-deploy-secret.yaml
apiVersion: v1
kind: Secret
metadata:
  name: "devops-deploy"
  labels:
    "jenkins.io/credentials-type": "usernamePassword"
    mycompany.com/jenkins: devops
---
# common-user-secret.yaml
apiVersion: v1
kind: Secret
metadata:
  name: "common-user"
  labels:
    "jenkins.io/credentials-type": "usernamePassword"
    mycompany.com/jenkins: common
---
# othergroup-user-secret.yaml
# This secret will NOT be in picked up by the devops jenkins instance
apiVersion: v1
kind: Secret
metadata:
  name: "othergroup-user"
  labels:
    "jenkins.io/credentials-type": "usernamePassword"
    mycompany.com/jenkins: othergroup
```

I couldn't think of a better way to configure this other than a system property.