Start secrets watcher before list operation
Because the watch operation no longer uses the resourceVersion
from the list operation, there is a small oppurtunity for the plugin
to miss some events. The watcher is now started first, then the
list operation is performed. Since the watcher keeps track of
its last resourceVersion internally this new ordering should the
providers secret map fully in sync.

Removed the JenkinsRule because it was causing startup issues
with the k8s secret provider plugin because it was not using
the test version with mocked k8s client. Switched to PowerMock
to mock the static ExtensionList.lookup method as a workaround
so we can test the administrative errors and loading secrets on
startup.